+++ changelog.md
@@ -12,6 +12,7 @@ Types of changes: Added, Changed, Deprecated, Removed, Fixed, Security
 - Method 'get_all' don't compatible with previous versions! Update method 'get_all' to support additional options. Options: filter, sort, limit, skip.
 - Add property 'id' to WorkerCollection.
 - Revalidate fields is enabled every time they are updated. The 'validate_assignment' parameter is set to True.
+- The default value of the field is now validated. The 'validate_default' parameter is set to True.
 
 
 ## [0.2.7] - 2024-07-10